requirements.txt
cachetools==5.3.0
deta==1.1.0
firebase_admin==6.1.0
python-dotenv==1.0.0
streamlit==1.22.0
streamlit_authenticator==0.1.5
